Static functions and common API function calls
Made all public function in AudioEngine static, including the Initialize and Terminate. There should only be one instance of the audio engine, so there is no need to hold a pointer to it.

Exposed the most common functions from the FMOD API and added Audio3DAttributes as an Alias to FMOD_3D_Attributes